examples/custom_model_example.py
"""
Example: How to create a custom diffusion model
自定义扩散模型示例

这个示例展示如何创建自己的扩散编辑模型实现
"""

import sys
from pathlib import Path
project_root = Path(__file__).parent.parent
sys.path.insert(0, str(project_root))

from PIL import Image
from src.models.diffusion.base_diffusion import BaseDiffusionModel


class MyCustomDiffusionModel(BaseDiffusionModel):
    """
    自定义扩散编辑模型示例
    
    步骤：
    1. 继承 BaseDiffusionModel
    2. 实现 _initialize() 方法
    3. 实现 edit_image() 方法
    4. （可选）实现 batch_edit() 方法以优化批处理
    """
    
    def _initialize(self):
        """
        初始化模型
        
        在这里加载你的模型权重、设置设备等
        """
        # 从配置中获取参数
        self.model_name = self.config.get("model_name", "my-model")
        self.device = self.config.get("device", "cuda")
        self.num_inference_steps = self.config.get("num_inference_steps", 50)
        self.guidance_scale = self.config.get("guidance_scale", 7.5)
        
        print(f"Initializing {self.model_name}...")
        
        # 加载你的实际模型
        # 例如使用 Hugging Face Diffusers:
        """
        from diffusers import StableDiffusionInstructPix2PixPipeline
        import torch
        
        self.pipe = StableDiffusionInstructPix2PixPipeline.from_pretrained(
            self.model_name,
            torch_dtype=torch.float16,
            safety_checker=None
        )
        self.pipe.to(self.device)
        self.pipe.enable_attention_slicing()
        """
        
        print("Model loaded successfully!")
    
    def edit_image(self, 
                   original_image: Image.Image,
                   edit_instruction: str,
                   **kwargs) -> Image.Image:
        """
        执行图像编辑
        
        Args:
            original_image: 原始PIL图像
            edit_instruction: 编辑指令文本
            **kwargs: 其他参数（会覆盖默认配置）
            
        Returns:
            编辑后的PIL图像
        """
        # 获取参数（优先使用kwargs，否则使用配置）
        num_steps = kwargs.get("num_inference_steps", self.num_inference_steps)
        guidance = kwargs.get("guidance_scale", self.guidance_scale)
        
        print(f"Editing with instruction: '{edit_instruction}'")
        
        # 调用实际的编辑模型
        """
        edited_image = self.pipe(
            prompt=edit_instruction,
            image=original_image,
            num_inference_steps=num_steps,
            image_guidance_scale=guidance,
        ).images[0]
        """
        
        # 示例：这里返回原图（实际使用时替换为真实编辑）
        edited_image = original_image.copy()
        
        return edited_image


# 类似地，创建自定义的Reward模型
from src.models.reward.base_reward import BaseRewardModel
from typing import Optional


class MyCustomRewardModel(BaseRewardModel):
    """
    自定义Reward评分模型示例
    
    步骤：
    1. 继承 BaseRewardModel
    2. 实现 _initialize() 方法
    3. 实现 score() 方法
    """
    
    def _initialize(self):
        """初始化Reward模型"""
        self.model_name = self.config.get("model_name", "my-reward-model")
        self.device = self.config.get("device", "cuda")
        
        print(f"Initializing reward model: {self.model_name}")
        
        # 加载你的实际模型
        # 例如使用 Vision-Language Model:
        """
        from transformers import AutoModel, AutoProcessor
        
        self.model = AutoModel.from_pretrained(self.model_name).to(self.device)
        self.processor = AutoProcessor.from_pretrained(self.model_name)
        """
        
        print("Reward model loaded!")
    
    def score(self,
              edited_image: Image.Image,
              original_description: str,
              edit_instruction: str,
              system_prompt: str,
              user_prompt: str,
              original_image: Optional[Image.Image] = None,
              **kwargs) -> float:
        """
        对编辑后的图像进行评分
        
        Returns:
            评分（0-10的浮点数）
        """
        print("Scoring edited image...")
        
        # 使用模型进行评分
        """
        # 准备输入
        inputs = self.processor(
            text=user_prompt,
            images=edited_image,
            return_tensors="pt"
        ).to(self.device)
        
        # 生成评分
        outputs = self.model.generate(**inputs, max_new_tokens=100)
        response = self.processor.decode(outputs[0], skip_special_tokens=True)
        
        # 从响应中提取分数
        score = self._extract_score(response)
        """
        
        # 示例：返回随机分数（实际使用时替换）
        import random
        score = random.uniform(6.0, 9.0)
        
        return score
    
    def _extract_score(self, response: str) -> float:
        """从模型响应中提取分数"""
        import re
        match = re.search(r'(\d+\.?\d*)', response)
        if match:
            score = float(match.group(1))
            return min(max(score, 0.0), 10.0)
        return 5.0


def main():
    """使用自定义模型的示例"""
    
    # 在配置文件中，指定你的自定义模型类路径：
    config_example = {
        "diffusion_model": {
            "class_path": "path.to.your.module.MyCustomDiffusionModel",
            "params": {
                "model_name": "timbrooks/instruct-pix2pix",
                "device": "cuda",
                "num_inference_steps": 50,
                "guidance_scale": 7.5
            }
        },
        "reward_model": {
            "class_path": "path.to.your.module.MyCustomRewardModel",
            "params": {
                "model_name": "your-vlm-model",
                "device": "cuda"
            }
        }
    }
    
    print("Custom model configuration example:")
    print(config_example)
    
    # 测试自定义模型
    print("\nTesting custom diffusion model:")
    diffusion_model = MyCustomDiffusionModel(config_example["diffusion_model"]["params"])
    
    # 创建测试图像
    test_image = Image.new("RGB", (512, 512), color="white")
    edited = diffusion_model.edit_image(test_image, "make it blue")
    print(f"Edited image size: {edited.size}")
    
    print("\nTesting custom reward model:")
    reward_model = MyCustomRewardModel(config_example["reward_model"]["params"])
    score = reward_model.score(
        edited_image=edited,
        original_description="A white image",
        edit_instruction="make it blue",
        system_prompt="You are an evaluator",
        user_prompt="Rate this image"
    )
    print(f"Score: {score:.2f}")


if __name__ == "__main__":
    main()


